refactor(entity-ui): extract Model & harness + Advanced into useModelHarness
Move the agent-template model/connection/harness/sandbox state and the
Model & harness + Advanced section JSX out of AgentTemplateControl into a
dedicated useModelHarness hook. The two sections share one coupled state
machine (the model/connection state feeds both), so they live together and
the hook returns the summaries + section bodies the orchestrator renders.

AgentTemplateControl drops from ~1.4k to 672 lines; behavior is unchanged.
